Remove numpy to make reward type consistent

+++ ple/games/utils/__init__.py
@@ -1,5 +1,2 @@
-import numpy as np
-
-
 def percent_round_int(percent, x):
-    return np.round(percent * x).astype(int)
+    return round(percent * x)
